@@ -2,14 +2,55 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <Version>0.6.0</Version>
-  </PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
 
+    <!-- Debug settings -->
+    <DebugType>portable</DebugType>
+    <DebugSymbols>true</DebugSymbols>
+
+    <!-- Auto-increment build version based on date/time -->
+    <Version>0.7.$([System.DateTime]::Now.DayOfYear).$([System.DateTime]::Now.Hour)</Version>
+    <AssemblyVersion>0.7.0.0</AssemblyVersion>
+    <FileVersion>0.7.0.0</FileVersion>
+    <InformationalVersion>0.7.$([System.DateTime]::Now.DayOfYear).$([System.DateTime]::Now.Hour)</InformationalVersion>
+
+    <!-- App info -->
+    <Authors>mstarczewski</Authors>
+    <Company>mstarczewski</Company>
+    <Product>RcloneFileWatcherCore</Product>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <RepositoryUrl>https://github.com/mstarczewski/RcloneFileWatcherCore</RepositoryUrl>
+
+    <!-- Deployment -->
+    <PublishSingleFile>true</PublishSingleFile>
+    <SelfContained>false</SelfContained>
+
+    <!-- General build -->
+    <Platforms>AnyCPU;x64</Platforms>
+    <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
+  </PropertyGroup>
+  
   <ItemGroup>
     <None Update="RcloneFileWatcherCoreConfig.cfg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
 
+  <!-- Generate version.json in output dir -->
+  <Target Name="GenerateVersionFile" AfterTargets="Build">
+    <WriteLinesToFile
+      File="$(OutputPath)version.json"
+      Lines='[
+  "Product: $(Product)",
+  "Version: $(Version)",
+  "FileVersion: $(FileVersion)",
+  "InformationalVersion: $(InformationalVersion)",
+  "Author: $(Authors)",
+  "Company: $(Company)",
+  "License: MIT",
+  "BuildDate: $([System.DateTime]::Now.ToString("yyyy-MM-dd HH:mm:ss"))"
+]'
+      Overwrite="true" />
+  </Target>
 </Project>
