@@ -2,11 +2,16 @@
 
 namespace RcloneFileWatcherCore.DTO
 {
-    class PathDTO
+    public class PathDTO
     {
         public string WatchingPath { get; set; }
         public string RcloneFilesFromPath { get; set; }
         public string RcloneBatch { get; set; }
         public List<string> ExcludeContains { get; set; }
+
+        public PathDTO()
+        {
+            ExcludeContains = new List<string>();
+        }
     }
 }
