@@ -2,11 +2,10 @@
 
 namespace RcloneFileWatcherCore.DTO
 {
-    class UpdateRcloneDTO
+    public class UpdateRcloneDTO
     {
         public bool Update { get; set; }
         public string RclonePath { get; set; }
-        public string RcloneWebsiteCurrentVersionAddress { get; set; }
-        public int ChceckUpdateHours { get; set; }
+        public int CheckUpdateHours { get; set; }
     }
 }
