@@ -6,43 +6,15 @@ namespace RcloneFileWatcherCore.Globals
     public static class TimeStamp
     {
         private static long _ticks = DateTime.Now.Ticks;
-        static ReaderWriterLockSlim readWriteLock = new ReaderWriterLockSlim();
-        public const int timeoutLock = 1000 * 10;
+
         public static void SetTimestampTicks()
         {
-            if (readWriteLock.TryEnterWriteLock(timeoutLock))
-            {
-                try
-                {
-                    _ticks = DateTime.Now.Ticks;
-                }
-                finally
-                {
-                    readWriteLock.ExitWriteLock();
-                }
-            }
-            else
-            {
-                _ticks = -1;
-            }
+            Interlocked.Exchange(ref _ticks, DateTime.Now.Ticks);
         }
+
         public static long GetTimestampTicks()
         {
-            if (readWriteLock.TryEnterReadLock(timeoutLock))
-            {
-                try
-                {
-                    return _ticks;
-                }
-                finally
-                {
-                    readWriteLock.ExitReadLock();
-                }
-            }
-            else
-            {
-                return -1;
-            }
+            return Interlocked.Read(ref _ticks);
         }
     }
 }
