@@ -1,16 +1,16 @@
 ﻿using RcloneFileWatcherCore.Logic.Interfaces;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using RcloneFileWatcherCore.DTO;
 using System.Text.Json;
 
 namespace RcloneFileWatcherCore.Logic
 {
-    class Config
+    internal class Config
     {
         private readonly string _configFileName;
         private readonly ILogger _logger;
+
         public Config(string configFileName, ILogger logger)
         {
             _configFileName = configFileName;
@@ -19,21 +19,27 @@ namespace RcloneFileWatcherCore.Logic
 
         internal ConfigDTO LoadConfig()
         {
+            if (!File.Exists(_configFileName))
+            {
+                _logger.WriteAlways($"Config file is missing: {_configFileName}");
+                return null;
+            }
+
             try
             {
-                if (!File.Exists(_configFileName))
+                using var stream = File.OpenRead(_configFileName);
+                var config = JsonSerializer.Deserialize<ConfigDTO>(stream);
+                if (config == null)
                 {
-                    _logger.WriteAlways("Config file is missing");
+                    _logger.WriteAlways("Config file is empty or invalid");
                     return null;
                 }
-
-                var _config = JsonSerializer.Deserialize<ConfigDTO>(File.ReadAllText(_configFileName));
-                _logger.Enable = _config.ConsoleWriter;
-                return _config;
+                _logger.Enable = config.ConsoleWriter;
+                return config;
             }
             catch (Exception ex)
             {
-                _logger.WriteAlways("Config file error");
+                _logger.WriteAlways($"Config file error: {_configFileName}");
                 _logger.WriteAlways(ex.ToString());
                 return null;
             }
